requirements.txt
python-tk
google-cloud-storage
flask
datetime
flask_sqlalchemy
sqlite3